raptor + orca + lemon + python: helpers-native migration (plumbing-only)
Four legacy runtime-1.0-era agents migrated to helpers-native
runtime-1.1.0 pattern. Same six-layer flip applied uniformly across
all four: mcporter subprocess → SenpiClient.mcp_call() HTTPS, scanner-
side create_position → producer push_signal() with LLM-gated runtime
action, openclaw cron → producer_daemon, Python state files for daily
caps → declarative risk.guard_rails, MARKET DSL exits →
FEE_OPTIMIZED_LIMIT (maker-first, taker fallback).

NO thesis changes. Each agent's unique scoring + gates preserved
verbatim.

ORCA v3.0 → v4.0.0:
- Gen-1 Vanilla Striker (FIRST_JUMP + base scoring + volume confirm)
- MIN_SCORE 9, 4-reason floor, MAX_LEVERAGE 7, MAX_POSITIONS 3
- 120-min cooldown, 90s tick cadence
- DSL preset preserved (45min hard_timeout, fast-cycle)

LEMON v1.3 → v2.0.0:
- Degen Fader on 12 crypto majors + 4 XYZ (BRENTOIL/CL/GOLD/SPX)
- MACRO_TREND_GATE (crypto only, |BTC 4h| > 3% blocks fades)
- MIN_SCORE 9, MARGIN_PCT 30%, MAX_LEVERAGE 10x
- Leverage tiers 5x/7x/10x, 5min tick cadence
- DSL preset preserved (480min hard_timeout, wide tiers for reversion)

PYTHON v1.2 → v2.0.0:
- The Patience Hunter — multi-day hold (96h target)
- 4h+1h+1d trend scoring, LONG bias bonus, MIN_SCORE 8
- 3x/5x/7x leverage tiers, conviction-scaled margin (25/30/40%)
- MACRO_TREND_GATE, SM hard-block on opposition
- 12h per-asset cooldown, 10min tick cadence
- DSL preset preserved (96h hard_timeout, weak_peak_cut DISABLED for
  patience thesis, Phase 1 retrace 30%)

RAPTOR v3.4 → v4.0.0:
- Hot Streak Follower — quality-first pipeline
- ELITE/RELIABLE weekly winners → strongest position → SM alignment
- Whale entry-discipline 5% threshold (v3.3, prevents buying their top)
- v3.4 nested-positions parser fix preserved
- MIN_SCORE 6, conviction-tier leverage 7x/8x/10x
- Per-trader event dedupe (4h), 3min tick cadence
- DSL preset preserved (360min hard_timeout, 5-tier Phase 2 ladder)

Each agent shares the same plumbing-flip pattern as the Dog/Mantis/
Bison/Condor migrations from the prior batch:
- 5-step install (openclaw.json registration → install runtime skill
  → pull files → env vars → recreate runtime + launch daemon)
- producer_daemon with wallet= + scanner= for /state alive_check
- Daemon must use the multi-path SDK probe pointing at
  ~/.openclaw/skills/senpi-trading-runtime/ first

Files deleted: orca-scanner.py, lemon-scanner.py, python-scanner.py,
raptor-scanner.py (all legacy v1-era full-agency scanners).
