+++ lemon/scripts/lemon_config.py
@@ -1,14 +1,15 @@
-"""LEMON v1.0 — Degen Fader Config Helper.
-Self-contained. Standard Senpi skill pattern.
+"""LEMON v2.0.0 — Shared config + MCP shim + helpers wrapper.
 
-v1.0.1: Fixed midnight rollover bug — load_trade_counter() now checks
-date before returning, preventing permanent lock at daily limit."""
+v2.0.0: senpi_runtime_helpers migration. mcporter_call now routes
+through SenpiClient.mcp_call(). _wrapper_client is exposed for
+push_signal access.
+"""
 # Copyright 2026 Senpi (https://senpi.ai)
 # Licensed under MIT
 
+import functools
 import json
 import os
-import subprocess
 import sys
 import tempfile
 import time
@@ -19,22 +20,57 @@ WORKSPACE = os.environ.get("OPENCLAW_WORKSPACE", "/data/workspace")
 SKILL_DIR = Path(WORKSPACE) / "skills" / "lemon-strategy"
 CONFIG_PATH = SKILL_DIR / "config" / "lemon-config.json"
 STATE_DIR = SKILL_DIR / "state"
+COOLDOWN_FILE = STATE_DIR / "asset-cooldowns.json"
 
 STATE_DIR.mkdir(parents=True, exist_ok=True)
 
 
+# ─── senpi_runtime_helpers (lazy + auth-validated) ───
+_sdk_candidates = [
+    str(Path.home() / ".openclaw" / "skills" / "senpi-trading-runtime"),
+    str(Path(os.environ.get("OPENCLAW_WORKSPACE", "/data/workspace")) / "skills" / "senpi-trading-runtime"),
+]
+_sdk_path = next(
+    (p for p in _sdk_candidates if (Path(p) / "senpi_runtime_helpers").is_dir()),
+    _sdk_candidates[0],
+)
+if _sdk_path not in sys.path:
+    sys.path.insert(0, _sdk_path)
+from senpi_runtime_helpers import SenpiClient, log_event  # type: ignore  # noqa: E402
+
+
+@functools.lru_cache(maxsize=1)
+def _get_wrapper_client() -> SenpiClient:
+    if not os.environ.get("SENPI_AUTH_TOKEN", "").strip():
+        raise RuntimeError(
+            "SENPI_AUTH_TOKEN is not set. Lemon's MCP calls and signal "
+            "POST both require it."
+        )
+    client = SenpiClient()
+    log_event("lemon_wrapper_enabled", sdk_path=_sdk_path)
+    return client
+
+
+class _WrapperClientProxy:
+    def __getattr__(self, name: str):
+        return getattr(_get_wrapper_client(), name)
+
+
+_wrapper_client = _WrapperClientProxy()
+
+
 def atomic_write(path, data):
     path = str(path)
-    d = os.path.dirname(path) or "."
-    os.makedirs(d, exist_ok=True)
-    fd, tmp = tempfile.mkstemp(dir=d, suffix=".tmp")
+    dir_name = os.path.dirname(path) or "."
+    os.makedirs(dir_name, exist_ok=True)
+    fd, tmp_path = tempfile.mkstemp(dir=dir_name, suffix=".tmp")
     try:
         with os.fdopen(fd, "w") as f:
-            json.dump(data, f, indent=2, default=str)
-        os.replace(tmp, path)
+            json.dump(data, f, indent=2)
+        os.replace(tmp_path, path)
     except BaseException:
         try:
-            os.unlink(tmp)
+            os.unlink(tmp_path)
         except OSError:
             pass
         raise
@@ -48,132 +84,38 @@ def load_config():
 
 
 def get_wallet_and_strategy():
-    w = os.environ.get("LEMON_WALLET", "")
-    s = os.environ.get("LEMON_STRATEGY_ID", "")
-    if not w or not s:
-        c = load_config()
-        w = w or c.get("wallet", "")
-        s = s or c.get("strategyId", "")
-    return w, s
+    wallet = os.environ.get("LEMON_WALLET", "").strip()
+    strategy_id = os.environ.get("LEMON_STRATEGY_ID", "").strip()
+    if not wallet or not strategy_id:
+        config = load_config()
+        wallet = wallet or config.get("wallet", "").strip()
+        strategy_id = strategy_id or config.get("strategyId", "").strip()
+    return wallet, strategy_id
 
 
-def load_trade_counter():
-    today = now_date()
-    p = STATE_DIR / "trade-counter.json"
-    default = {"date": today, "entries": 0, "dailyLoss": 0,
-               "consecutiveLosses": 0, "gate": "OPEN", "last_entry_ts": 0}
-    if p.exists():
-        try:
-            with open(p) as f:
-                tc = json.load(f)
-            if tc.get("date") != today:
-                return {"date": today, "entries": 0, "dailyLoss": 0,
-                        "consecutiveLosses": 0, "gate": "OPEN", "last_entry_ts": 0}
-            for k, v in default.items():
-                if k not in tc:
-                    tc[k] = v
-            return tc
-        except (json.JSONDecodeError, IOError):
-            pass
-    return dict(default)
-
-
-def save_trade_counter(tc):
-    if tc.get("date") != now_date():
-        tc = {"date": now_date(), "entries": 0, "dailyLoss": 0,
-              "consecutiveLosses": 0, "gate": "OPEN", "last_entry_ts": 0}
-    atomic_write(str(STATE_DIR / "trade-counter.json"), tc)
-
-
-def is_on_cooldown(coin, minutes=180):
-    p = STATE_DIR / "cooldowns.json"
-    if not p.exists():
-        return False
+def mcp_call(tool, **params):
     try:
-        with open(p) as f:
-            cooldowns = json.load(f)
-    except (json.JSONDecodeError, IOError):
-        return False
-    entry = cooldowns.get(coin)
-    if not entry:
-        return False
-    return time.time() < entry.get("until", 0)
+        return _wrapper_client.mcp_call(tool, **params)
+    except Exception as e:  # noqa: BLE001
+        log_event("lemon_mcp_call_failed", tool=tool, error=str(e))
+        return None
 
 
-def set_cooldown(coin, minutes=180):
-    p = STATE_DIR / "cooldowns.json"
-    cooldowns = {}
-    if p.exists():
-        try:
-            with open(p) as f:
-                cooldowns = json.load(f)
-        except (json.JSONDecodeError, IOError):
-            pass
-    cooldowns[coin] = {"until": time.time() + minutes * 60, "set_at": now_iso()}
-    atomic_write(str(p), cooldowns)
+mcporter_call = mcp_call
 
 
-def load_scan_history():
-    p = STATE_DIR / "scan-history.json"
-    if p.exists():
-        try:
-            with open(p) as f:
-                return json.load(f)
-        except (json.JSONDecodeError, IOError):
-            pass
-    return {"scans": []}
-
-
-def save_scan_history(history):
-    scans = history.get("scans", [])
-    if len(scans) > 30:
-        history["scans"] = scans[-30:]
-    atomic_write(str(STATE_DIR / "scan-history.json"), history)
-
-
-def mcporter_call(tool, retries=2, timeout=30, **params):
-    args = json.dumps(params) if params else "{}"
-    cmd = ["mcporter", "call", "senpi", tool, "--args", args]
-    for attempt in range(retries):
-        try:
-            r = subprocess.run(cmd, capture_output=True, text=True, timeout=timeout)
-            if r.returncode != 0:
-                if attempt < retries - 1:
-                    time.sleep(2)
-                    continue
-                return None
-            raw = json.loads(r.stdout)
-            if isinstance(raw, dict) and "content" in raw:
-                content = raw["content"]
-                if isinstance(content, list) and content:
-                    first = content[0]
-                    if isinstance(first, dict) and "text" in first:
-                        try:
-                            return json.loads(first["text"])
-                        except (json.JSONDecodeError, TypeError):
-                            pass
-            return raw
-        except subprocess.TimeoutExpired:
-            if attempt < retries - 1:
-                time.sleep(2)
-                continue
-            return None
-        except (json.JSONDecodeError, Exception):
-            return None
-    return None
-
-
-def get_positions(wallet=None):
+def get_clearinghouse(wallet):
     if not wallet:
-        wallet, _ = get_wallet_and_strategy()
-    if not wallet:
-        return 0, []
-    ch = mcporter_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
-    if not ch or not isinstance(ch, dict):
+        return None
+    return mcp_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
+
+
+def get_positions(wallet):
+    ch = get_clearinghouse(wallet)
+    if not ch:
         return 0, []
     data = ch.get("data", ch)
-    positions = []
-    account_value = 0
+    positions, account_value = [], 0
     for section in ("main", "xyz"):
         s = data.get(section, {})
         if not isinstance(s, dict):
@@ -188,35 +130,37 @@ def get_positions(wallet=None):
             positions.append({
                 "coin": pos.get("coin", ""),
                 "direction": "LONG" if szi > 0 else "SHORT",
-                "szi": szi,
-                "size": abs(szi),
+                "upnl": float(pos.get("unrealizedPnl", 0)),
                 "margin": float(pos.get("marginUsed", 0)),
                 "entryPrice": float(pos.get("entryPx", 0)),
-                "markPrice": float(pos.get("markPx", 0)),
-                "leverage": float(
-                    pos.get("leverage", {}).get("value", 5)
-                    if isinstance(pos.get("leverage"), dict)
-                    else pos.get("leverage", 5)
-                ),
-                "upnl": float(pos.get("unrealizedPnl", 0)),
+                "size": abs(szi),
             })
     return account_value, positions
 
 
+def is_asset_cooled_down(asset, cooldown_minutes=120):
+    if not COOLDOWN_FILE.exists():
+        return False
+    try:
+        with open(COOLDOWN_FILE) as f:
+            cooldowns = json.load(f)
+    except (json.JSONDecodeError, IOError):
+        return False
+    if asset not in cooldowns:
+        return False
+    last_ts = cooldowns[asset].get("ts", 0)
+    elapsed_min = (time.time() - last_ts) / 60
+    return elapsed_min < cooldown_minutes
+
+
 def output(data):
-    print(json.dumps(data, default=str))
+    print(json.dumps(data))
     sys.stdout.flush()
 
 
 def log(msg):
-    print(f"[LEMON] {msg}", file=sys.stderr)
+    print(f"[lemon-v2] {msg}", file=sys.stderr, flush=True)
 
-
-def now_ts():
-    return time.time()
 
 def now_iso():
     return datetime.now(timezone.utc).isoformat()
-
-def now_date():
-    return datetime.now(timezone.utc).strftime("%Y-%m-%d")
