+++ orca/runtime.yaml
@@ -1,21 +1,60 @@
 name: orca-tracker
 version: 1.0.0
 description: >
-  Gen-2 Striker — FIRST_JUMP detection enhanced with momentum event
-  quality confirmation. Only enters when quality traders (ELITE/RELIABLE)
-  are driving the move. Fast-cycling DSL.
+  ORCA v4.0.0 — Gen-1 Vanilla Striker, helpers-native rewrite.
+  Plumbing-only migration from v3.0. NO thesis change. FIRST_JUMP /
+  IMMEDIATE_MOVER + base Striker scoring + volume confirmation
+  preserved verbatim. Producer (orca-producer.py) emits signals via
+  push_signal() to runtime /signals; runtime opens via LLM-gated
+  orca_entry action with FEE_OPTIMIZED_LIMIT. risk.guard_rails
+  enforces daily caps + per-asset cooldown + drawdown halt.
 
 strategy:
   wallet: "${WALLET_ADDRESS}"
-  budget: 1000
   slots: 3
   margin_per_slot: 180
+  trading_risk: moderate
   enabled: true
+
+risk:
+  data_retention_hours: 72
+  guard_rails:
+    daily_loss_limit_pct: 15
+    max_entries_per_day: 6
+    bypass_max_entries_per_day_on_profit: false
+    max_consecutive_losses: 3
+    cooldown_minutes: 30
+    drawdown_halt_pct: 25
+    drawdown_reset_on_day_rollover: false
+    per_asset_cooldown_minutes: 120
 
 scanners:
   - name: position_tracker
     type: position_tracker
     interval: 10s
+
+  - name: orca_signals
+    type: external_scanner
+    outputs:
+      signals: true
+      context: false
+    config:
+      fields:
+        score: { type: number, required: true }
+        leverage: { type: number, required: true }
+        marginUsd: { type: number, required: true }
+        mode: { type: string, required: false }
+        currentRank: { type: number, required: false }
+        rankJump: { type: number, required: false }
+        isFirstJump: { type: boolean, required: false }
+        isContribExplosion: { type: boolean, required: false }
+        contribVelocity: { type: number, required: false }
+        volRatio: { type: number, required: false }
+        contribution: { type: number, required: false }
+        traders: { type: number, required: false }
+        priceChg4h: { type: number, required: false }
+        reasons: { type: string, required: false }
+        heldAssets: { type: array, required: false }
 
 actions:
   - name: position_tracker_action
@@ -23,9 +62,72 @@ actions:
     decision_mode: rule
     scanners: [position_tracker]
 
+  - name: orca_entry
+    action_type: OPEN_POSITION
+    decision_mode: llm
+    decision_model: "${ORCA_DECISION_MODEL}"
+    scanners: [orca_signals]
+    min_confidence: 7
+    params:
+      order_type: FEE_OPTIMIZED_LIMIT
+      fee_optimized_limit_options:
+        ensure_execution_as_taker: true
+        execution_timeout_seconds: 60
+    context:
+      - type: signal
+        scanner: orca_signals
+    decision_prompt: |
+      You are Orca's pass-through gate. The producer has already applied
+      every filter (FIRST_JUMP / IMMEDIATE_MOVER detection, 4h trend
+      alignment, 15m velocity freshness, rank jump >= 15, score >= 9
+      with 4+ reasons, volume >= 1.5x). Your job is to honor the signal
+      unless something is structurally broken.
+
+      SIGNAL:
+      {{signal_orca_signals}}
+
+      ## STRICT OUTPUT RULES
+
+      1. `asset` MUST be copied EXACTLY from signal.asset.
+      2. `direction` MUST match signal.direction exactly.
+      3. `leverage` MUST be copied from signal.data.leverage exactly (typically 7).
+      4. `marginUsd` MUST be copied from signal.data.marginUsd exactly.
+
+      ## HARD SKIPS (output execute: false)
+
+      - score < 9 (defensive — producer filters)
+      - leverage <= 0 OR marginUsd <= 0 (malformed)
+      - mode != "STRIKER" (only Striker mode is valid)
+
+      Minimum confidence to execute is 7. Honor unless structurally broken.
+
+      ## OUTPUT
+
+      {
+        "execute": true OR false,
+        "actionType": "OPEN_POSITION",
+        "confidence": 7,
+        "reasoning": "one-liner citing score + rankJump + 4h_alignment",
+        "payload": {
+          "signals": [
+            {
+              "asset": "copy signal.asset EXACTLY",
+              "direction": "copy signal.direction EXACTLY",
+              "leverage": "copy signal.data.leverage EXACTLY",
+              "marginUsd": "copy signal.data.marginUsd EXACTLY",
+              "reason": "orca striker entry"
+            }
+          ]
+        }
+      }
+
 exit:
   engine: dsl
   interval_seconds: 30
+  order_type: FEE_OPTIMIZED_LIMIT
+  fee_optimized_limit_options:
+    ensure_execution_as_taker: true
+    execution_timeout_seconds: 60
   dsl_preset:
     hard_timeout:
       enabled: true
@@ -52,3 +154,6 @@ exit:
 
 notifications:
   telegram_chat_id: "${TELEGRAM_CHAT_ID}"
+  dsl_lifecycle: true
+  dsl_notify_sl_updates: false
+  action_lifecycle: true
