+++ python/runtime.yaml
@@ -1,28 +1,57 @@
 name: python-tracker
-version: 1.2.0
+version: 1.0.0
 description: >
-  PYTHON v1.2 — disable weak_peak_cut (completes patience thesis fix).
-  v1.1 loosened Phase 1 retrace 10→30 and extended time cuts after
-  Python's self-diagnostic showed trades averaging <4h despite a 2-4
-  day thesis target. v1.2 removes weak_peak_cut entirely — a
-  multi-day patience thesis cannot tolerate ANY time-based cut. A
-  position peaking at 2% ROE on day 1 of a 3-day thesis can still
-  develop into a monster over day 3. Phase 1 retrace (30%) + max_loss
-  (20%) + Phase 2 tiers own all exits. dead_weight_cut + hard_timeout
-  (96h) preserved — 96h matches thesis target as an outer bound.
-  Scanner unchanged. v1.1 retrace/dead_weight loosening preserved.
+  PYTHON v2.0.0 — The Patience Hunter, helpers-native rewrite.
+  Plumbing-only migration from v1.2. NO thesis change. Top 50 HL
+  assets, multi-day hold thesis (96h hard_timeout), LONG-biased,
+  3-7x leverage tiers, MIN_SCORE 8, MACRO_TREND_GATE, conviction-
+  scaled margin (25%/30%/40%). DSL preset preserved from v1.2 with
+  weak_peak_cut DISABLED.
 
 strategy:
   wallet: "${WALLET_ADDRESS}"
-  budget: 1000
   slots: 2
   margin_per_slot: 250
+  trading_risk: moderate
   enabled: true
+
+risk:
+  data_retention_hours: 72
+  guard_rails:
+    daily_loss_limit_pct: 15
+    max_entries_per_day: 3
+    bypass_max_entries_per_day_on_profit: false
+    max_consecutive_losses: 3
+    cooldown_minutes: 120
+    drawdown_halt_pct: 25
+    drawdown_reset_on_day_rollover: false
+    per_asset_cooldown_minutes: 720       # 12h — preserved from v1.x ASSET_COOLDOWN
 
 scanners:
   - name: position_tracker
     type: position_tracker
     interval: 10s
+
+  - name: python_signals
+    type: external_scanner
+    outputs:
+      signals: true
+      context: false
+    config:
+      fields:
+        score: { type: number, required: true }
+        leverage: { type: number, required: true }
+        marginUsd: { type: number, required: true }
+        reasons: { type: array, required: false }
+        trend4h: { type: string, required: false }
+        trend1h: { type: string, required: false }
+        mom1h: { type: number, required: false }
+        mom4h: { type: number, required: false }
+        mom1d: { type: number, required: false }
+        funding: { type: number, required: false }
+        rsi: { type: number, required: false }
+        volRatio: { type: number, required: false }
+        heldAssets: { type: array, required: false }
 
 actions:
   - name: position_tracker_action
@@ -30,38 +59,98 @@ actions:
     decision_mode: rule
     scanners: [position_tracker]
 
-# ─── EXIT (DSL) — Patience profile (v1.1 aligned with thesis) ──
-# Python's thesis: 2-4 day holds, high-conviction signals, let winners
-# run 100%+. v1.0's DSL was a scalper's leash.
-# v1.1 loosens the trailing-retrace floor 3x so normal intraday noise
-# doesn't trigger exits, and extends time cuts to match multi-day hold
-# expectations. Phase 1 max_loss preserved as the absolute backstop.
+  - name: python_entry
+    action_type: OPEN_POSITION
+    decision_mode: llm
+    decision_model: "${PYTHON_DECISION_MODEL}"
+    scanners: [python_signals]
+    min_confidence: 7
+    params:
+      order_type: FEE_OPTIMIZED_LIMIT
+      fee_optimized_limit_options:
+        ensure_execution_as_taker: false
+        execution_timeout_seconds: 30
+    context:
+      - type: signal
+        scanner: python_signals
+    decision_prompt: |
+      You are Python's pass-through gate. Python is a PATIENCE HUNTER —
+      it holds for DAYS (96h target). The producer has applied every
+      filter (4h trend, 1h match, MACRO_TREND_GATE, MIN_SCORE 8, SM
+      hard-block, conviction-tier sizing). Honor the signal unless
+      structurally broken.
+
+      SIGNAL:
+      {{signal_python_signals}}
+
+      ## STRICT OUTPUT RULES
+
+      1. `asset` MUST be copied EXACTLY from signal.asset.
+      2. `direction` MUST match signal.direction exactly (LONG-biased).
+      3. `leverage` from signal.data.leverage exactly (3/5/7).
+      4. `marginUsd` from signal.data.marginUsd exactly (25%/30%/40%).
+
+      ## HARD SKIPS
+
+      - score < 8 (defensive)
+      - leverage <= 0 OR marginUsd <= 0
+      - heldAssets contains this asset
+
+      Minimum confidence to execute is 7. Honor unless structurally broken.
+
+      ## OUTPUT
+
+      {
+        "execute": true OR false,
+        "actionType": "OPEN_POSITION",
+        "confidence": integer 1-10,
+        "reasoning": "concrete sentence citing score + trend4h + mom1d + sm_aligned",
+        "payload": {
+          "signals": [
+            {
+              "asset": "copy signal.asset EXACTLY",
+              "direction": "copy signal.direction EXACTLY",
+              "leverage": "copy signal.data.leverage EXACTLY",
+              "marginUsd": "copy signal.data.marginUsd EXACTLY",
+              "reason": "PYTHON_PATIENCE ASSET DIR"
+            }
+          ]
+        }
+      }
+
 exit:
   engine: dsl
   interval_seconds: 30
+  order_type: FEE_OPTIMIZED_LIMIT
+  fee_optimized_limit_options:
+    ensure_execution_as_taker: true
+    execution_timeout_seconds: 60
   dsl_preset:
     hard_timeout:
       enabled: true
-      interval_in_minutes: 5760       # 96h = 4 days (matches thesis target)
+      interval_in_minutes: 5760        # 96h = 4 days (preserved from v1.2)
     weak_peak_cut:
-      enabled: false                   # v1.2: DISABLED. Multi-day patience thesis cannot tolerate ANY time-based cut. Phase 1 retrace (30%) + max_loss (20%) + Phase 2 tiers own exits.
+      enabled: false                   # v1.2 DISABLED — multi-day patience thesis
       interval_in_minutes: 2880
       min_value: 3.0
     dead_weight_cut:
       enabled: true
-      interval_in_minutes: 1440        # v1.1: 360 → 1440 (24h no movement)
+      interval_in_minutes: 1440        # 24h no movement
     phase1:
       enabled: true
-      max_loss_pct: 20.0               # absolute hard stop (unchanged)
-      retrace_threshold: 30            # v1.1: 10 → 30 (core thesis-alignment fix)
+      max_loss_pct: 20.0
+      retrace_threshold: 30
     phase2:
       enabled: true
       tiers:
-        - { trigger_pct: 5,   lock_hw_pct: 15 }   # LOOSE initial lock — let winners breathe
+        - { trigger_pct: 5,   lock_hw_pct: 15 }
         - { trigger_pct: 12,  lock_hw_pct: 35 }
         - { trigger_pct: 25,  lock_hw_pct: 60 }
-        - { trigger_pct: 50,  lock_hw_pct: 80 }   # pr0br000's MON/WLD-tier
-        - { trigger_pct: 100, lock_hw_pct: 94 }   # v1.1: was 90; bumped to final tier
+        - { trigger_pct: 50,  lock_hw_pct: 80 }
+        - { trigger_pct: 100, lock_hw_pct: 94 }
 
 notifications:
   telegram_chat_id: "${TELEGRAM_CHAT_ID}"
+  dsl_lifecycle: true
+  dsl_notify_sl_updates: false
+  action_lifecycle: true
