+++ python/scripts/python_config.py
@@ -1,16 +1,14 @@
-"""PYTHON Strategy — Shared config, MCP helpers, state I/O.
+"""PYTHON v2.0.0 — Shared config + MCP shim + helpers wrapper.
 
-Fleet-standard helper module for python-scanner.py.
-The Patience Hunter — multi-day holds, wide universe, low leverage,
-derived from pr0br000's Arena Week 2-3 winning pattern.
+v2.0.0: senpi_runtime_helpers migration. mcporter_call now routes
+through SenpiClient.mcp_call().
 """
 # Copyright 2026 Senpi (https://senpi.ai)
 # Licensed under MIT
-# Source: https://github.com/Senpi-ai/senpi-skills
 
+import functools
 import json
 import os
-import subprocess
 import sys
 import tempfile
 import time
@@ -26,7 +24,38 @@ COOLDOWN_FILE = STATE_DIR / "asset-cooldowns.json"
 STATE_DIR.mkdir(parents=True, exist_ok=True)
 
 
-# ─── Atomic Write ────────────────────────────────────────────
+_sdk_candidates = [
+    str(Path.home() / ".openclaw" / "skills" / "senpi-trading-runtime"),
+    str(Path(os.environ.get("OPENCLAW_WORKSPACE", "/data/workspace")) / "skills" / "senpi-trading-runtime"),
+]
+_sdk_path = next(
+    (p for p in _sdk_candidates if (Path(p) / "senpi_runtime_helpers").is_dir()),
+    _sdk_candidates[0],
+)
+if _sdk_path not in sys.path:
+    sys.path.insert(0, _sdk_path)
+from senpi_runtime_helpers import SenpiClient, log_event  # type: ignore  # noqa: E402
+
+
+@functools.lru_cache(maxsize=1)
+def _get_wrapper_client() -> SenpiClient:
+    if not os.environ.get("SENPI_AUTH_TOKEN", "").strip():
+        raise RuntimeError(
+            "SENPI_AUTH_TOKEN is not set. Python's MCP calls and signal "
+            "POST both require it."
+        )
+    client = SenpiClient()
+    log_event("python_wrapper_enabled", sdk_path=_sdk_path)
+    return client
+
+
+class _WrapperClientProxy:
+    def __getattr__(self, name: str):
+        return getattr(_get_wrapper_client(), name)
+
+
+_wrapper_client = _WrapperClientProxy()
+
 
 def atomic_write(path, data):
     path = str(path)
@@ -35,7 +64,7 @@ def atomic_write(path, data):
     fd, tmp_path = tempfile.mkstemp(dir=dir_name, suffix=".tmp")
     try:
         with os.fdopen(fd, "w") as f:
-            json.dump(data, f, indent=2, default=str)
+            json.dump(data, f, indent=2)
         os.replace(tmp_path, path)
     except BaseException:
         try:
@@ -45,8 +74,6 @@ def atomic_write(path, data):
         raise
 
 
-# ─── Config ──────────────────────────────────────────────────
-
 def load_config():
     if CONFIG_PATH.exists():
         with open(CONFIG_PATH) as f:
@@ -55,141 +82,30 @@ def load_config():
 
 
 def get_wallet_and_strategy():
-    wallet = os.environ.get("PYTHON_WALLET", "")
-    strategy_id = os.environ.get("PYTHON_STRATEGY_ID", "")
+    wallet = os.environ.get("PYTHON_WALLET", "").strip()
+    strategy_id = os.environ.get("PYTHON_STRATEGY_ID", "").strip()
     if not wallet or not strategy_id:
         config = load_config()
-        wallet = wallet or config.get("wallet", "")
-        strategy_id = strategy_id or config.get("strategyId", "")
+        wallet = wallet or config.get("wallet", "").strip()
+        strategy_id = strategy_id or config.get("strategyId", "").strip()
     return wallet, strategy_id
 
 
-# ─── State I/O ───────────────────────────────────────────────
-
-def load_state(filename="python-state.json"):
-    path = STATE_DIR / filename
-    if path.exists():
-        try:
-            with open(path) as f:
-                return json.load(f)
-        except (json.JSONDecodeError, IOError):
-            return {}
-    return {}
+def mcp_call(tool, **params):
+    try:
+        return _wrapper_client.mcp_call(tool, **params)
+    except Exception as e:  # noqa: BLE001
+        log_event("python_mcp_call_failed", tool=tool, error=str(e))
+        return None
 
 
-def save_state(data, filename="python-state.json"):
-    atomic_write(str(STATE_DIR / filename), data)
-
-
-# ─── Trade Counter ───────────────────────────────────────────
-
-def load_trade_counter():
-    today = now_date()
-    path = STATE_DIR / "trade-counter.json"
-    default = {
-        "date": today, "entries": 0, "realizedPnl": 0,
-        "gate": "OPEN", "gateReason": None,
-        "lastResults": [],
-        "last_entry_ts": 0,
-        "positionsOpenedByCoin": {},  # coin → last open ts (per-asset cooldown)
-    }
-    if path.exists():
-        try:
-            with open(path) as f:
-                tc = json.load(f)
-            if tc.get("date") != today:
-                tc["date"] = today
-                tc["entries"] = 0
-                tc["realizedPnl"] = 0
-                tc["lastResults"] = []
-            for k, v in default.items():
-                if k not in tc:
-                    tc[k] = v
-            return tc
-        except (json.JSONDecodeError, IOError):
-            pass
-    return dict(default)
-
-
-def save_trade_counter(tc):
-    tc["updatedAt"] = now_iso()
-    atomic_write(str(STATE_DIR / "trade-counter.json"), tc)
-
-
-# ─── Asset Cooldowns (per-asset, after exit) ─────────────────
-
-def load_cooldowns():
-    if COOLDOWN_FILE.exists():
-        try:
-            with open(COOLDOWN_FILE) as f:
-                return json.load(f)
-        except (json.JSONDecodeError, IOError):
-            return {}
-    return {}
-
-
-def save_cooldowns(cooldowns):
-    atomic_write(str(COOLDOWN_FILE), cooldowns)
-
-
-def is_asset_cooled_down(token, cooldown_minutes=720):
-    """Patience Hunter uses 12h default cooldown — we don't re-enter quickly."""
-    cooldowns = load_cooldowns()
-    if token not in cooldowns:
-        return False
-    exit_ts = cooldowns[token].get("exitTimestamp", 0)
-    elapsed_min = (now_ts() - exit_ts) / 60
-    return elapsed_min < cooldown_minutes
-
-
-def set_asset_cooldown(token, reason="position_exit"):
-    cooldowns = load_cooldowns()
-    cooldowns[token] = {
-        "exitTimestamp": now_ts(),
-        "reason": reason,
-        "setAt": now_iso(),
-    }
-    save_cooldowns(cooldowns)
-
-
-# ─── MCP Helpers ─────────────────────────────────────────────
-
-def mcporter_call(tool, retries=2, timeout=30, **params):
-    args = json.dumps(params) if params else "{}"
-    cmd = ["mcporter", "call", "senpi", tool, "--args", args]
-    for attempt in range(retries):
-        try:
-            r = subprocess.run(cmd, capture_output=True, text=True, timeout=timeout)
-            if r.returncode != 0:
-                if attempt < retries - 1:
-                    time.sleep(2)
-                    continue
-                return None
-            raw = json.loads(r.stdout)
-            if isinstance(raw, dict) and "content" in raw:
-                content = raw["content"]
-                if isinstance(content, list) and content:
-                    first = content[0]
-                    if isinstance(first, dict) and "text" in first:
-                        try:
-                            return json.loads(first["text"])
-                        except (json.JSONDecodeError, TypeError):
-                            pass
-            return raw
-        except subprocess.TimeoutExpired:
-            if attempt < retries - 1:
-                time.sleep(2)
-                continue
-            return None
-        except (json.JSONDecodeError, Exception):
-            return None
-    return None
+mcporter_call = mcp_call
 
 
 def get_clearinghouse(wallet):
     if not wallet:
         return None
-    return mcporter_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
+    return mcp_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
 
 
 def get_positions(wallet):
@@ -220,22 +136,29 @@ def get_positions(wallet):
     return account_value, positions
 
 
+def is_asset_cooled_down(asset, cooldown_minutes=720):
+    if not COOLDOWN_FILE.exists():
+        return False
+    try:
+        with open(COOLDOWN_FILE) as f:
+            cooldowns = json.load(f)
+    except (json.JSONDecodeError, IOError):
+        return False
+    if asset not in cooldowns:
+        return False
+    last_ts = cooldowns[asset].get("ts", 0)
+    elapsed_min = (time.time() - last_ts) / 60
+    return elapsed_min < cooldown_minutes
+
+
 def output(data):
-    print(json.dumps(data, default=str))
+    print(json.dumps(data))
     sys.stdout.flush()
 
 
 def log(msg):
-    print(f"[PYTHON] {msg}", file=sys.stderr)
-
-
-def now_ts():
-    return time.time()
+    print(f"[python-v2] {msg}", file=sys.stderr, flush=True)
 
 
 def now_iso():
     return datetime.now(timezone.utc).isoformat()
-
-
-def now_date():
-    return datetime.now(timezone.utc).strftime("%Y-%m-%d")
