+++ raptor/README.md
@@ -1,49 +1,90 @@
-# Raptor — Hot Streak Follower
+# 🦖 Raptor v4.0.0 — Hot Streak Follower (senpi_runtime_helpers)
 
-**Runtime:** 1.0  ·  **Asset:** Multi-asset  ·  **Version:** 3.3.0
+Part of [Senpi Trading Skills](https://github.com/Senpi-ai/senpi-skills).
 
-## Thesis
+**Plumbing-only migration from v3.4. NO thesis change.** v3.4 quality-first pipeline (ELITE/RELIABLE weekly winners → strongest position → SM alignment), whale entry-discipline 5% threshold, nested-positions parser, MIN_SCORE 6, conviction-tier leverage all preserved verbatim.
 
-RAPTOR v3.3 — Hot Streak Follower (entry-price discipline tightened).
+## Install
 
-See `SKILL.md` and `runtime.yaml` for full scoring components, gates, and DSL configuration.
+### Step 0 — `openclaw.json` plugin registration (one-time per host)
 
-## Scoring components
-
-Defined in the producer/scanner. See `runtime.yaml` `scanner:` section and the producer script in this folder for the current scoring weights and gates.
-
-## Entry / Exit
-
-- **Entry:** Producer-emitted signals scored against MIN_SCORE gate.
-- **Exit:** DSL (Dynamic Stop-Loss) Phase 1 + Phase 2 trailing exits per `runtime.yaml` `dsl:` section.
-- **Time cuts:** See `dsl:` config in runtime.yaml.
-
-## Fleet rules applied
-
-- Standard fleet drawdown gate.
-- Fee budget tracking via shared infra.
-- Producer reentrancy guard via fcntl lockfile.
-
-## Configuration
-
-Operator-specific values configured at deploy time.
-
-## Recent version notes
-
-```
-RAPTOR v3.3 — Hot Streak Follower (entry-price discipline tightened).
-v3.2 introduced whale-entry-price gate at 20% threshold. Raptor
-self-diagnosed 2026-04-23: the gate was still too loose. Scanner was
-re-firing on whales already deep in profit, taking all downside risk
-with none of their PnL cushion. v3.3 tightens
-MAX_PRICE_RUN_PCT_FROM_WHALE_ENTRY from 20% to 5% — 4x tighter, still
-leaves room for "whale in early profit" piggyback. Raptor's proposed
-1-2% would likely zero the scanner; 5% is the first tightening with
-room for further per data. Scanner-only change. DSL unchanged.
+```json
+{
+  "plugins": {
+    "entries": {
+      "runtime": {
+        "enabled": true,
+        "config": {
+          "stateDir": "/data/.openclaw/senpi-state",
+          "apiKey": "<your SENPI_AUTH_TOKEN>",
+          "autoUpdate": { "enabled": false }
+        }
+      }
+    }
+  }
+}
 ```
 
-## Related
+`openclaw gateway restart`. Confirm with `curl -s -m 5 http://127.0.0.1:8787/state | head -c 200`.
 
-- Top-level repo README: `../README.md`
-- Runtime spec: `../senpi-trading-runtime/`
-- DSL plugin: `../dsl-dynamic-stop-loss/`
+### Step 1 — senpi-trading-runtime skill
+
+```bash
+npx skills add https://github.com/Senpi-ai/senpi-skills --skill senpi-trading-runtime -g -y
+which senpi-helpers
+```
+
+### Step 2 — Pull Raptor v4.0.0
+
+```bash
+mkdir -p /data/workspace/skills/raptor-strategy/{config,scripts,state}
+for f in scripts/raptor-producer.py scripts/raptor_config.py runtime.yaml SKILL.md README.md \
+         config/raptor-config.json; do
+  curl -fsSL "https://raw.githubusercontent.com/Senpi-ai/senpi-skills/main/raptor/$f" \
+    -o "/data/workspace/skills/raptor-strategy/$f"
+done
+```
+
+### Step 3 — Configure wallet + chat ID
+
+Edit `/data/workspace/skills/raptor-strategy/config/raptor-config.json` with `wallet`, `strategyId`, `chatId`.
+
+### Step 4 — Env vars
+
+```bash
+export RAPTOR_WALLET=<your-raptor-wallet>
+export SENPI_AUTH_TOKEN=...
+export RAPTOR_DECISION_MODEL=gemini-2.5-pro
+```
+
+### Step 5 — Recreate runtime + launch daemon
+
+```bash
+openclaw senpi runtime list | grep raptor
+openclaw senpi runtime delete <old-raptor-runtime-id>
+openclaw senpi runtime create --path /data/workspace/skills/raptor-strategy/runtime.yaml
+
+openclaw cron list | grep raptor
+openclaw cron delete <raptor-cron-id>
+
+nohup python3 -u /data/workspace/skills/raptor-strategy/scripts/raptor-producer.py \
+  > /tmp/raptor-producer.log 2>&1 &
+```
+
+If `daemon_aborted_no_runtime: alive_check returned False`, re-register the runtime.
+
+## Smoke test
+
+```bash
+tail -f /tmp/raptor-producer.log | jq -c 'select(.event=="daemon_tick_finished")' | head -3
+```
+
+Expect `status=ok` every 3 min.
+
+## Thesis (preserved from v3.4)
+
+Find ELITE/RELIABLE traders currently winning weekly. Pick their strongest position by |delta_pnl|. Confirm SM alignment + 4h/1h price agreement. **Apply whale entry-discipline (5% threshold)** — if the asset has already run >5% in the whale's favor from their entry, skip; we'd be buying their top. Bonus +1/+2 points if we'd get a better fill than the whale.
+
+## License
+
+MIT — Built by Senpi (https://senpi.ai).
