+++ raptor/scripts/raptor_config.py
@@ -1,15 +1,14 @@
-"""RAPTOR Strategy — Shared config, MCP helpers, state I/O.
+"""RAPTOR v4.0.0 — Shared config + MCP shim + helpers wrapper.
 
-v3.0: added to repo (was previously missing — agents had local copies only).
-Matches fleet-standard pattern from phoenix_config.py / wolverine_config.py.
+v4.0.0: senpi_runtime_helpers migration. mcporter_call now routes
+through SenpiClient.mcp_call().
 """
 # Copyright 2026 Senpi (https://senpi.ai)
 # Licensed under MIT
-# Source: https://github.com/Senpi-ai/senpi-skills
 
+import functools
 import json
 import os
-import subprocess
 import sys
 import tempfile
 import time
@@ -24,17 +23,47 @@ STATE_DIR = SKILL_DIR / "state"
 STATE_DIR.mkdir(parents=True, exist_ok=True)
 
 
-# ─── Atomic Write ────────────────────────────────────────────
+_sdk_candidates = [
+    str(Path.home() / ".openclaw" / "skills" / "senpi-trading-runtime"),
+    str(Path(os.environ.get("OPENCLAW_WORKSPACE", "/data/workspace")) / "skills" / "senpi-trading-runtime"),
+]
+_sdk_path = next(
+    (p for p in _sdk_candidates if (Path(p) / "senpi_runtime_helpers").is_dir()),
+    _sdk_candidates[0],
+)
+if _sdk_path not in sys.path:
+    sys.path.insert(0, _sdk_path)
+from senpi_runtime_helpers import SenpiClient, log_event  # type: ignore  # noqa: E402
+
+
+@functools.lru_cache(maxsize=1)
+def _get_wrapper_client() -> SenpiClient:
+    if not os.environ.get("SENPI_AUTH_TOKEN", "").strip():
+        raise RuntimeError(
+            "SENPI_AUTH_TOKEN is not set. Raptor's MCP calls and signal "
+            "POST both require it."
+        )
+    client = SenpiClient()
+    log_event("raptor_wrapper_enabled", sdk_path=_sdk_path)
+    return client
+
+
+class _WrapperClientProxy:
+    def __getattr__(self, name: str):
+        return getattr(_get_wrapper_client(), name)
+
+
+_wrapper_client = _WrapperClientProxy()
+
 
 def atomic_write(path, data):
-    """Write JSON atomically via tmp file + os.replace."""
     path = str(path)
     dir_name = os.path.dirname(path) or "."
     os.makedirs(dir_name, exist_ok=True)
     fd, tmp_path = tempfile.mkstemp(dir=dir_name, suffix=".tmp")
     try:
         with os.fdopen(fd, "w") as f:
-            json.dump(data, f, indent=2, default=str)
+            json.dump(data, f, indent=2)
         os.replace(tmp_path, path)
     except BaseException:
         try:
@@ -44,8 +73,6 @@ def atomic_write(path, data):
         raise
 
 
-# ─── Config ──────────────────────────────────────────────────
-
 def load_config():
     if CONFIG_PATH.exists():
         with open(CONFIG_PATH) as f:
@@ -54,79 +81,38 @@ def load_config():
 
 
 def get_wallet_and_strategy():
-    wallet = os.environ.get("RAPTOR_WALLET", "")
-    strategy_id = os.environ.get("RAPTOR_STRATEGY_ID", "")
+    wallet = os.environ.get("RAPTOR_WALLET", "").strip()
+    strategy_id = os.environ.get("RAPTOR_STRATEGY_ID", "").strip()
     if not wallet or not strategy_id:
         config = load_config()
-        wallet = wallet or config.get("wallet", "")
-        strategy_id = strategy_id or config.get("strategyId", "")
+        wallet = wallet or config.get("wallet", "").strip()
+        strategy_id = strategy_id or config.get("strategyId", "").strip()
     return wallet, strategy_id
 
 
-# ─── State I/O ───────────────────────────────────────────────
-
-def load_state(filename="state.json"):
-    path = STATE_DIR / filename
-    if path.exists():
-        try:
-            with open(path) as f:
-                return json.load(f)
-        except (json.JSONDecodeError, IOError):
-            pass
-    return {}
+def mcp_call(tool, **params):
+    try:
+        return _wrapper_client.mcp_call(tool, **params)
+    except Exception as e:  # noqa: BLE001
+        log_event("raptor_mcp_call_failed", tool=tool, error=str(e))
+        return None
 
 
-def save_state(data, filename="state.json"):
-    atomic_write(str(STATE_DIR / filename), data)
+mcporter_call = mcp_call
 
 
-# ─── MCP Helpers ─────────────────────────────────────────────
-
-def mcporter_call(tool, retries=2, timeout=30, **params):
-    """Call a Senpi MCP tool via mcporter. Array syntax, no shell=True."""
-    args = json.dumps(params) if params else "{}"
-    cmd = ["mcporter", "call", "senpi", tool, "--args", args]
-    for attempt in range(retries):
-        try:
-            r = subprocess.run(cmd, capture_output=True, text=True, timeout=timeout)
-            if r.returncode != 0:
-                if attempt < retries - 1:
-                    time.sleep(2)
-                    continue
-                return None
-            raw = json.loads(r.stdout)
-            if isinstance(raw, dict) and "content" in raw:
-                content = raw["content"]
-                if isinstance(content, list) and content:
-                    first = content[0]
-                    if isinstance(first, dict) and "text" in first:
-                        try:
-                            return json.loads(first["text"])
-                        except (json.JSONDecodeError, TypeError):
-                            pass
-            return raw
-        except subprocess.TimeoutExpired:
-            if attempt < retries - 1:
-                time.sleep(2)
-                continue
-            return None
-        except (json.JSONDecodeError, Exception):
-            return None
-    return None
-
-
-def get_positions(wallet=None):
-    """Returns (account_value, positions_list)."""
+def get_clearinghouse(wallet):
     if not wallet:
-        wallet, _ = get_wallet_and_strategy()
-    if not wallet:
-        return 0, []
-    ch = mcporter_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
-    if not ch or not isinstance(ch, dict):
+        return None
+    return mcp_call("strategy_get_clearinghouse_state", strategy_wallet=wallet)
+
+
+def get_positions(wallet):
+    ch = get_clearinghouse(wallet)
+    if not ch:
         return 0, []
     data = ch.get("data", ch)
-    positions = []
-    account_value = 0
+    positions, account_value = [], 0
     for section in ("main", "xyz"):
         s = data.get(section, {})
         if not isinstance(s, dict):
@@ -141,32 +127,21 @@ def get_positions(wallet=None):
             positions.append({
                 "coin": pos.get("coin", ""),
                 "direction": "LONG" if szi > 0 else "SHORT",
-                "szi": szi,
-                "size": abs(szi),
+                "upnl": float(pos.get("unrealizedPnl", 0)),
                 "margin": float(pos.get("marginUsed", 0)),
                 "entryPrice": float(pos.get("entryPx", 0)),
-                "markPrice": float(pos.get("markPx", 0)),
-                "leverage": float(
-                    pos.get("leverage", {}).get("value", 5)
-                    if isinstance(pos.get("leverage"), dict)
-                    else pos.get("leverage", 5)
-                ),
-                "upnl": float(pos.get("unrealizedPnl", 0)),
+                "size": abs(szi),
             })
     return account_value, positions
 
 
 def output(data):
-    print(json.dumps(data, default=str))
+    print(json.dumps(data))
     sys.stdout.flush()
 
 
 def log(msg):
-    print(f"[RAPTOR] {msg}", file=sys.stderr)
-
-
-def now_ts():
-    return time.time()
+    print(f"[raptor-v4] {msg}", file=sys.stderr, flush=True)
 
 
 def now_iso():
